refact: unified reading of moderator
